vhost: fix slot index calculation in async vhost
When the packet receiving failure and the DMA ring full occur
simultaneously in the asynchronous vhost, the slot_idx needs to be
reduced by 1. For packed virtqueue, the slot index should be
ring_size - 1, if the slot_idx is currently 0, since the ring size is
not necessarily the power of 2.

Signed-off-by: Cheng Jiang <cheng1.jiang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>