vdpa/ifc: set vring state callback update data path
To support multi queue, in the case that first queue is ready
and device is configured, when more queues need to be configured,
we just close and restart data path.
This also fix the situation that using set_vring_state callback
to disable one queue will cause all vfio interrupts being disabled.

Signed-off-by: Andy Pei <andy.pei@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>