usertools/telemetry: add new telemetry client
A lot of DPDK apps do not provide feedback to their users while
running, but many interesting metrics are available through the
app's telemetry socket. Currently, the main way to view these
metrics is to query them using the dpdk-telemetry.py script.
This is useful to check a few values or to check a value once.
However, it is difficult to observe key values over time and it
can be difficult to correctly parse JSON in one’s head!

It is proposed to add a new Python tool to usertools that will
provide our users with a real-time, easy and visually appealing
way to view statistics without leaving their terminals.
The tool can provide various key statistics about any DPDK app
in real-time. Various real-time graphs are available (using plotext),
which makes visualizing the metrics easier.

The tool runs completely in terminal and is responsive to any
terminal size down to 80 characters.
The TUI is provided using the rich Python module.

The tool is already being used by some of Intel’s DPDK developers
internally and we feel that it could be useful for the wider
DPDK community.

A talk about this new tool will be presented at DPDK userspace.

Screenshot: https://cwalsh.tech/dpdk/tui2.png

Signed-off-by: Conor Walsh <conor.walsh@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>