vhost: compilation fix for GCC-12
GCC-12 complains about the possible use of un-initialized array. At
compile time it seems like it is not able to evaluate the size as it
involves run-time variable and at compile time it seems like gcc assumes
value of "size" variable to be zero which makes gcc-12 to jump the while
loop.
"size = pkt->pkt_len + sizeof(struct virtio_net_hdr_mrg_rxbuf);"

As part of the fix, "while (){}" is replaced by "do {} while()" which make
the compiler to generate a code in which buf_vec will never be used
un-initialized.

../lib/vhost/virtio_net.c:941:35: error:
		'buf_vec[0].buf_len' may be used uninitialized
		[-Werror=maybe-uninitialized]
  941 |         buf_len = buf_vec[vec_idx].buf_len;
      |                   ~~~~~~~~~~~~~~~~^~~~~~~~
../lib/vhost/virtio_net.c: In function 'virtio_dev_rx_packed':
../lib/vhost/virtio_net.c:1285:27: note: 'buf_vec' declared here
 1285 |         struct buf_vector buf_vec[BUF_VECTOR_MAX];
      |                           ^~~~~~~
cc1: all warnings being treated as errors

Fixes: 93520085efda ("vhost: add packed ring single enqueue")
Cc: stable@dpdk.org

Signed-off-by: Amit Prakash Shukla <amitprakashs@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>