vhost: introduce VHOST_USER_RESET_VRING
To support the reset operation for an individual virtqueue, we
introduce a new message VHOST_USER_RESET_VRING. When the feature
VIRTIO_F_RING_RESET feature has been successfully negotiated, This
message is submitted by the front-end to reset an individual
virtqueue to initial states in the back-end. The reply is needed
to ensure that the reset operation is complete.

Signed-off-by: Kangjie Xu <kangjie.xu@linux.alibaba.com>
Signed-off-by: Xuan Zhuo <xuanzhuo@linux.alibaba.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>