lib: move mbuf next pointer to first cache line
Swapped position of mbuf next pointer and second dynamic field (dynfield2)
if the build is configured to use IOVA as VA. This is to move the mbuf
next pointer to first cache line. kni library is disabled for this
change as it depends on the offset value of next pointer.

Signed-off-by: Shijith Thotton <sthotton@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>