eal: fix thread names for high order lcores
In this patch we suggest a new name for lcore-worker.
In case of higher order (greater than 99) logical cores, name is truncated
(length is restricted to 16 characters, including the
terminating null byte ('\0')) and it makes hard to follow threads.

Ex: This issue can be generated using following arguments:
	--lcores=0,10@1,100@2
Then we have;
lcore-worker-10
lcore-worker-10

Signed-off-by: Abdullah Ömer Yamaç <omer.yamac@ceng.metu.edu.tr>

Acked-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>