drivers/bus: set device NUMA node to unknown by default
The dev->device.numa_node field is set by each bus driver for
every device it manages to indicate on which NUMA node this device lies.

When this information is unknown, the assigned value is not consistent
across the bus drivers.

Set the default value to SOCKET_ID_ANY (-1) by all bus drivers
when the NUMA information is unavailable.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>