drivers: suggestion on meson without version file
Most of the drivers don't have a special version.map file. They just
included due to the compilation issue and needs to be updated for each
release.

These version.map files include:
DPDK_23 {
  local: *;
};

In this patch, we removed the necessity of the version files and
you don't need to update these files for each release, you can just
remove them.

Signed-off-by: Abdullah Ömer Yamaç <omer.yamac@ceng.metu.edu.tr>
Suggested-by: Ferruh Yigit <ferruh.yigit@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>