net/iavf: fix error of virtchnl command
When the device is bonded, bond pmd will register callback for LSC event.
This callback will execute some virtchnl commands in eal-intr-thread to
reinitialize the device with interrupts disabled. In this case, responses
to all commands not be received.

This commit starts a thread to handle all events to fix this issue.

Fixes: 48de41ca11f0 ("net/avf: enable link status update")
CC: stable@dpdk.org

Signed-off-by: Yiding Zhou <yidingx.zhou@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>